refactor(cli): split Clap schema by topic (Phase 9)
Implements Phase 9 of the Rust structure refactor roadmap. Splits the
1170-LOC `src/cli.rs` into three topic files plus a thin `mod.rs`,
giving each command family its own home while preserving the full
Clap schema and every existing `jackin::cli::*` import path.

Split:

- `mod.rs` (34 LOC, was 1170) — `HELP_STYLES` + `BANNER` constants,
  submodule declarations, and re-exports of `Cli`, `Command`,
  `WorkspaceCommand`, `ConfigCommand`, `AuthCommand`, `MountCommand`,
  `TrustCommand`.
- `root.rs` (413 LOC) — `Cli`, `Command` (top-level dispatch with
  `Load`, `Hardline`, `Eject`, `Exile`, `Purge`, `Launch`,
  `Workspace`, `Config`).
- `workspace.rs` (425 LOC) — `WorkspaceCommand` (Create, List, Show,
  Edit, Prune, Remove).
- `config.rs` (378 LOC) — `ConfigCommand`, `AuthCommand`,
  `MountCommand`, `TrustCommand`.

Verbatim moves. No flag renames. No help-text drift. No variant
reordering. No attribute edits. A Clap schema change would be
user-visible behavior, which the plan explicitly forbids.

Test Preservation Policy: all 47 tests in `src/cli.rs`'s test block
migrated verbatim, co-located with the enum they exercise:

- `root.rs` — 17 tests (parses_load_*, parses_hardline_*,
  parses_launch_*, plus top-level help-output tests).
- `workspace.rs` — 17 tests (parses_workspace_* and workspace help
  tests).
- `config.rs` — 13 tests (parses_config_* and config help tests).
- `mod.rs` — 0 tests; it's glue only.

The small `strip_ansi` / `help_text` test helpers were copied
verbatim into each of `workspace.rs` and `config.rs` that needed
them — same bytes, not rewritten. Tests themselves unchanged.

Visibility changes (scoped to the `cli` module):
- `HELP_STYLES`: private → `pub(super)` so `root.rs::Command::Load`
  can reference it.
- `BANNER`: private → `pub(super)` for the same reason.

No public API widened. All 12 external `cli::*` references in
`src/main.rs` and `src/app/mod.rs` resolve unchanged.

Verification per COMMITS.md:
- cargo fmt -- --check — pass
- cargo clippy — zero warnings
- cargo nextest run — 418 tests run: 418 passed, 0 skipped (unchanged)
